split reporting classes in report and report utils
