+++ testar/src/org/testar/protocols/DesktopProtocol.java
@@ -43,6 +43,7 @@ import org.testar.monkey.alayer.exceptions.StateBuildException;
 import org.testar.monkey.ConfigTags;
 import org.testar.OutputStructure;
 import org.testar.managers.InputDataManager;
+import org.testar.reporting_proofofconcept.ReportManager;
 
 import java.io.File;
 import java.io.IOException;
@@ -56,6 +57,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
     protected static double SCROLL_ARROW_SIZE = 36; // sliding arrows
     protected static double SCROLL_THICK = 16; //scroll thickness
     protected Reporting htmlReport;
+    private ReportManager reportManager;
     protected State latestState;
 
     /**
@@ -65,6 +67,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
     protected void preSequencePreparations() {
         //initializing the HTML sequence report:
         htmlReport = getReporter();
+        reportManager = new ReportManager((mode() == Modes.Replay), true, false);
     }
     
     /**
@@ -99,6 +102,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
     	latestState = super.getState(system);
         //adding state to the HTML sequence report:
         htmlReport.addState(latestState);
+        reportManager.addState(latestState);
         return latestState;
     }
 
@@ -133,6 +137,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
     	actions = super.preSelectAction(system, state, actions);
     	// adding available actions into the HTML report:
     	htmlReport.addActions(actions);
+        reportManager.addActions(actions);
     	return actions;
     }
 
@@ -147,6 +152,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
     protected boolean executeAction(SUT system, State state, Action action){
         // adding the action that is going to be executed into HTML report:
         htmlReport.addSelectedAction(state, action);
+        reportManager.addSelectedAction(state, action);
         return super.executeAction(system, state, action);
     }
     
@@ -157,6 +163,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
     protected boolean replayAction(SUT system, State state, Action action, double actionWaitTime, double actionDuration){
         // adding the action that is going to be executed into HTML report:
         htmlReport.addSelectedAction(state, action);
+        reportManager.addSelectedAction(state, action);
         return super.replayAction(system, state, action, actionWaitTime, actionDuration);
     }
 
@@ -170,11 +177,13 @@ public class DesktopProtocol extends GenericUtilsProtocol {
 
         if(mode() == Modes.Replay) {
             htmlReport.addTestVerdict(getReplayVerdict().join(processVerdict));
+            reportManager.addTestVerdict(getReplayVerdict().join(processVerdict));
             status = (getReplayVerdict().join(processVerdict)).verdictSeverityTitle();
             statusInfo = (getReplayVerdict().join(processVerdict)).info();
         }
         else {
             htmlReport.addTestVerdict(getFinalVerdict());
+            reportManager.addTestVerdict(getFinalVerdict());
             status = (getFinalVerdict()).verdictSeverityTitle();
             statusInfo = (getFinalVerdict()).info();
         }
@@ -194,6 +203,7 @@ public class DesktopProtocol extends GenericUtilsProtocol {
                 + " " + status + " \"" + statusInfo + "\"" );
 
         htmlReport.close();
+        reportManager.generateReport();
     }
 
     /**
